feat(cli): maker CLI created with all non-schema commands
