+++ src/cli/compile.js
@@ -1,0 +1,1 @@
+export default () => {};
